requirements.txt
Django==4.2.1
coverage==7.2.7
tox==4.5.2
pytest==7.3.1
pytest-django==4.5.2
pytest-xdist==3.3.1
pytest-cov==4.1.0
pytest-playwright==0.3.3
playwright==1.34.0
black==23.3.0
ruff==0.0.270
build==0.10.0
twine==4.0.2
factory_boy==3.2.1